% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3generics.R, R/utils.R
\name{as_draws}
\alias{as_draws}
\alias{as_draws.rjuliabugs}
\alias{as_draws.array}
\title{Convert to posterior::draws format}
\usage{
as_draws(x, ...)

\method{as_draws}{rjuliabugs}(x, ...)

\method{as_draws}{array}(x, ...)
}
\arguments{
\item{x}{An object to convert (e.g., a \code{rjuliabugs} object or a 3D array).}

\item{...}{Further arguments passed to specific methods.}
}
\value{
A \code{draws} object (e.g., \code{draws_array}) or a modified \code{rjuliabugs} object.
}
\description{
Generic function to convert an object to a \code{\link[posterior:draws]{posterior::draws}} representation
(e.g., \code{draws_array}, \code{draws_matrix}, or other formats).
}
\seealso{
\code{\link[posterior:draws]{posterior::as_draws()}}
}
