% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3generics.R, R/utils.R
\name{as_mcmc}
\alias{as_mcmc}
\alias{as_mcmc.rjuliabugs}
\alias{as_mcmc.array}
\title{Convert to coda::mcmc or mcmc.list format}
\usage{
as_mcmc(x, ...)

\method{as_mcmc}{rjuliabugs}(x, ...)

\method{as_mcmc}{array}(x, ...)
}
\arguments{
\item{x}{An object to convert (e.g., a `rjuliabugs` object or a 3D array).}

\item{...}{Further arguments passed to specific methods.}
}
\value{
An object of class `mcmc`, `mcmc.list`, or a `rjuliabugs` object with updated `params`.
}
\description{
Generic function to convert MCMC output to [`coda::mcmc`] or [`coda::mcmc.list`] format.
}
\seealso{
[coda::as.mcmc()], [coda::mcmc()]
}
