% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_sampler_is_defined}
\alias{check_sampler_is_defined}
\title{Ensure Unique Name for Julia Sampler (Internal)}
\usage{
check_sampler_is_defined(name)
}
\arguments{
\item{name}{A character string representing a Julia object name.}
}
\value{
A unique name not currently defined in Julia.
}
\description{
Checks whether a Julia variable name is already defined in the \code{Main} module.
If the name exists, appends random characters until a unique name is found.
}
\keyword{internal}
