% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delete_julia_obj}
\alias{delete_julia_obj}
\title{Delete an object from the Julia Main environment}
\usage{
delete_julia_obj(obj_name)
}
\arguments{
\item{obj_name}{A character string specifying the name of the Julia object to be deleted.
This should correspond to a variable or symbol previously defined in the Julia \code{Main} module.}
}
\description{
This function removes a variable or object from the Julia \code{Main} module using JuliaCall.
It is useful for cleaning up or resetting objects defined in the Julia environment from R.
}
\examples{
\dontrun{
JuliaCall::julia_command("x = 10")  # Define a Julia variable
delete_julia_obj("x")               # Delete it
}

}
