% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_params}
\alias{get_params}
\title{Extract and convert posterior parameters}
\usage{
get_params(rjuliabugs, params, posterior_type = "array")
}
\arguments{
\item{rjuliabugs}{An object containing posterior samples and a `name` attribute used to extract parameters.}

\item{params}{A character vector specifying which parameters to extract. **(NOTE: This argument is unused in the current function body.)**}

\item{posterior_type}{A character string specifying the desired output format. Options are: \code{"array"}, \code{"rvar"}, \code{"mcmc"}, \code{"draws"}.}
}
\value{
The posterior samples converted to the specified format. The return type depends on \code{posterior_type}:
\item{\code{"array"}}{A 3D array of posterior samples.}
\item{\code{"rvar"}}{An object of class \code{rvar} from the \pkg{posterior} package.}
\item{\code{"mcmc"}}{An \code{mcmc} or \code{mcmc.list} object from the \pkg{coda} package.}
\item{\code{"draws"}}{An object of class \code{draws_array} or similar from the \pkg{posterior} package.}
}
\description{
This function extracts parameters from an object created by `rjuliabugs` and converts them to a specified posterior format.
}
\details{
The function assumes that \code{get_params_from_name_raw()} and other referenced variables such as \code{params_to_save} and \code{n_chain} are available in the current environment.
}
\examples{
\dontrun{
get_params(rjuliabugs = fit, params = c("alpha", "beta"), posterior_type = "array")
}
}
\keyword{internal}
