% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3generics.R, R/utils.R
\name{as_draws}
\alias{as_draws}
\alias{as_draws.rjuliabugs}
\alias{as_draws.array}
\title{Convert to posterior::draws format}
\usage{
as_draws(x, ...)

\method{as_draws}{rjuliabugs}(x, ...)

\method{as_draws}{array}(x, ...)
}
\arguments{
\item{x}{An object to convert (e.g., a `rjuliabugs` object or a 3D array).}

\item{...}{Further arguments passed to specific methods.}
}
\value{
A `draws` object (e.g., `draws_array`) or a modified `rjuliabugs` object.
}
\description{
Generic function to convert an object to a [`posterior::draws`] representation
(e.g., `draws_array`, `draws_matrix`, or other formats).
}
\seealso{
[posterior::as_draws()]
}
