% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3generics.R, R/utils.R
\name{as_mcmc}
\alias{as_mcmc}
\alias{as_mcmc.rjuliabugs}
\alias{as_mcmc.array}
\title{Convert to coda::mcmc or mcmc.list format}
\usage{
as_mcmc(x, ...)

\method{as_mcmc}{rjuliabugs}(x, ...)

\method{as_mcmc}{array}(x, ...)
}
\arguments{
\item{x}{An object to convert (e.g., a \code{rjuliabugs} object or a 3D array).}

\item{...}{Further arguments passed to specific methods.}
}
\value{
An object of class \code{mcmc}, \code{mcmc.list}, or a \code{rjuliabugs} object with updated \code{params}.
}
\description{
Generic function to convert Markov Chain Monte Carlo (MCMC) output to \code{\link[coda:mcmc]{coda::mcmc}} or \code{\link[coda:mcmc.list]{coda::mcmc.list}} format.
}
\seealso{
\code{\link[coda:mcmc]{coda::as.mcmc()}}, \code{\link[coda:mcmc]{coda::mcmc()}}
}
