% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3generics.R, R/utils.R
\name{as_rvar}
\alias{as_rvar}
\alias{as_rvar.rjuliabugs}
\alias{as_rvar.array}
\title{Convert to posterior::rvar format}
\usage{
as_rvar(x, ...)

\method{as_rvar}{rjuliabugs}(x, ...)

\method{as_rvar}{array}(x, n_mcmc = NULL, ...)
}
\arguments{
\item{x}{An object to convert (e.g., a \code{rjuliabugs} object or a 3D numeric array).}

\item{...}{Further arguments passed to specific methods.}

\item{n_mcmc}{(For arrays only) Number of Markov Chain Monte Carlo (MCMC) chains. Required if \code{x} is an array.}
}
\value{
An object of class \code{rvar}, or an updated \code{rjuliabugs} object with \code{params} converted to \code{rvar}.
}
\description{
Generic function to convert an object to a \code{\link[posterior:rvar]{posterior::rvar}} representation.
This is typically used to convert Markov Chain Monte Carlo (MCMC) output into a more flexible and vectorized format.
}
\seealso{
\code{\link[posterior:rvar]{posterior::rvar()}}
}
